requirements.txt
accelerate==1.2.1
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
appscript==1.3.0
attrs==24.3.0
byaldi==0.0.7
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.8
colpali-engine==0.3.5
configobj==5.0.9
configparser==7.1.0
datasets==3.2.0
dill==0.3.8
docx2pdf==0.1.8
etelemetry==0.3.1
eval-type-backport==0.2.2
fastapi==0.115.6
filelock==3.16.1
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.9.0
gputil==1.4.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.0
idna==3.10
importlib-resources==6.4.5
isodate==0.6.1
itsdangerous==2.2.0
jinja2==3.1.5
joblib==1.4.2
jsonlines==4.0.0
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.5.0
mpmath==1.3.0
mteb==1.6.35
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
peft==0.11.1
pillow==10.4.0
propcache==0.2.1
prov==2.0.1
psutil==6.1.1
puremagic==1.28
pyarrow==18.1.0
pydantic==2.10.4
pydantic-core==2.27.2
pydot==3.0.3
pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
pypdf2==3.0.1
pyproject-hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytils==0.4.1
pytrec-eval-terrier==0.5.6
pytz==2024.2
pyxnat==1.6.2
pyyaml==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
ray==2.39.0
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
ruff==0.7.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==2.2.1
sentencepiece==0.2.0
setuptools==75.5.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.41.3
sympy==1.13.1
tenacity==8.5.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.3
tools==0.1.9
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
traits==6.4.3
transformers==4.46.2
typer==0.13.0
typing-extensions==4.12.2
typing-inspect==0.9.0
tzdata==2024.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
uuid7==0.1.0
uvicorn==0.32.1
uvloop==0.21.0
vllm==0.6.4.post1
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wrapt==1.16.0
xmltodict==0.13.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.50
zipp==3.21.0
